base: robot: Restore Lua semantics for setControllerInput (hopefully)
The semantics for this function in Lua were
- `_controllerInput = nil` means "reuse the last trajectory"
- `_controllerInput = {}` means "halt the robot"
- `_controllerInput.spline != nil` means "use this spline"
- Calling `setControllerInput` twice is an error, except if the robot
  was either previously set to halt, or is now set to halt
Furthermore, there are the `v_f`/... fields used by Hidden, but lets
ignore them for now.

The first implementation in Typescript compared `_controllerInput === {}`.
This has different semantics than in Lua though as Javascript compares
objects by reference. Lua compares them by value. This was later changed
by 3c93b020a1. However, that commit had a different semantic than the
Lua version.

This commit hopefully restores Lua semantics. Furthermore, it is
considered a type error to set both a spline and the v_f/... fields. It
is unclear to me whether this is actually disallowed by Amun, however
callers of setControllerInput use it this way.