glados: g/support: recompute zones more often
Previously the supporter zones where only changed when the number of supporters
changed, which was enough when the zones always looked the same for a specific
number of robots, but now with more adaptable zones this breaks.
Now the zones also get recomputed if the main attacker leaves its zone or the
objective changes.