glados: Introduce dummy robots
Even if we don't like it: the reality is we don't always have 11
fully functional robots. However, most robots are at least able
to drive. This commit introduces dummy robots: Their only job is
to be inside the opponent half, not break any rules and hopefully
be annoying.

Inside the Robot config in Ra you can set flags for whether a
Robot can dribble and shoot. If any of those flags is set to
false, the robot is put into the dummy PoolGroup. The Dummy
group assignes zones to each dummy robot. The dummy behavior
avoids the ball and moves back to the own half if the rules
require so.