eslint: move package.json and node_modules to root directory
Previously the node_modules packages were installed under strategy/typescript,
but this directory is supposed to only contain typescript and javascript files
directly related to the strategy.
Therefore the tooling for the linter was moved to the root directory of this
repository.
This has the added benefit, that it is easier to exclude the node_modules from
compilation.