tsconfig: remove default typeRoots which included unnecessary files
Also remove the hack in the setup.sh that removed the first few bytes of a
eslint plugin, because the typescript compiler could not parse it.
The excluded files are now really excluded, but more importantly if eslint
is installed the node_modules directory is now not parsed anymore, which should
not only avoid weird errors, but also improve compile times.