update package-lock.json using older version of npm
The previous package-lock.json was created with a version of npm that
was newer than the one in the CI which caused a weird error message and made
nmp ci fail.
This was not a huge issue, because the install_typescript_language_tools.sh
already used npm install as a fallback if npm ci failed, but it cluttered
the CI output showing errors that are not really errors that need to be
fixed.