glados: a/a/feintpass: Make feintSamplings hysteresis easier to read
The previous implementation relied on the fact that a series of if, else if, else if
checks would never run into the else block if the hysteresis variable was set to zero.
This is rather hard to read and confusing. The new implementation just checks if there
was a well-defined previous decision from the last frame and if a hysteresis would be
necessary because the passDist is close to one of the threshold values. Then the
previous decision is kept. Otherwise the decision is made with hard cuts.