glados: feintpass: Introduces group/feintpass
The feintpass group is introduced for two reasons:
1. Now that Dummys can do feintpass, two robots may want to
feint the same position. The group acts as a central instance
which decides which robot actually gets to do the feint.
(Dummys are preferred, since Supporters can do other usefull stuff)

2. Since the group continuously tells the robots through messaging
that they should do feintpass, it is actually possible to reconstruct
all neccessary internal state of the behavior, if it is killed e.g.
when the MA changes (see #994).

Additionally feintPassTargets are introduced, which can be used to
uniquely identify a feintPosition at a given Pass, even if the details
of the pass itself (timing, attackPosition, ballPos) change over time.