amun: Implement automatic entrypoints and accompanying GUI
Automatic entrypoints are designed to be loaded when the game stage
transitions to a specific stage, with the exception of FIRST_HALF (and
FIRST_HALF_PRE). Detecting a change to the first_half stage is
challenging due to its use as a default value. However this is
inconsequential as we already select an entrypoint manually prior to the
start of the game.

The primary purpose of this feature is to enable the automatic loading
of the lineup entrypoint when the break period begins.