glados: feintpass: FeintPos does not need to be reached without endspeed
The following situation caused problems:
Robot A feints pass of Robot B to Robot C.
Robot C plans to pass to A.
Robot A goes into AcceptPass.
Robot C redecides to pass to different Robot D.
In the following frame A checks if it can still reach the FeintPos of
the now imminent pass from B to C (it is still very close).
FeintPos cannot be reached anymore without endSpeed in time.
Therefore A does not continue FeintPass and does BreakPass instead.

In general, it is true that for the FeintPass timing it is not
important that the robot arrives at the FeintPos without residual endSpeed.
It is in principle imaginable for the feinting Robot to just 'drive through'
the passline.