glados: feintpass: passInfos should not be compared by reference
In the feintsamplings hysteresis, the passInfos are used as keys in the lastFeintSamplings map.
Therefore, if we check lastFeintSamplings.has(), the equality is checked by reference.
This is a bad idea, since the reference to the passInfo object might no longer be the same,
even if it is essentially still the same pass.

Therefore instead is is now checked, if
a) The MA is the same robot, by clearing lastFeintSamplings everytime the MA changes
and
b) The target of the passInfo is the same
under these conditions, two passes are considered to be the same and therefore a hysteresis is applied.