glados: trainer/defense.ts: check all defended opponents instead of only head
Previously only head of this._messaging.receive(MessageType.defendedOpponent)
was checked, but it is very much possible that there are multiple robots in
this list, even though it won't happen often.
The list is also unordered, so it does not make sense to only check the head
of the list.
Either check all defended robots or don't check at all, but checking all seems
more helpful here.