glados: t/a/support: Move Constructor Params to SupportParameters
Within Support, the typesytem now mandates that manualDefaultPos and
manualPassDest must be both set or unset. This is preferred over
automatically computing passDest if only manualDefaultPos is given, as
the calculating of passDest in Support lacks essential context.
Therefore, it is now necessary for the caller to provide both parameters
or none at all.

Fixes #841