marvin: te/t/volley: do not chip the pass
This (kind of) seems to revert 4a56756d0778d6cc34b79023cffaf7ab9e9c5027.
It does no such thing.
4a56756d0778d6cc34b79023cffaf7ab9e9c5027 introduced a new parameter to t/pass,
the then second parameter, that can be used to tell t/pass 'not to wait'.

This parameter told t/pass not to listen to messages to delay the pass
to accomoate the receiving robot. Presuambly because this test task
does not work with messages in the same way a normal game does,
this paramter was necessary.

932cf1822857d4e5ca0fe71b75e73ca1749e9207 rewrote - half a year later - almost all of t/pass.
One of the changes makes t/pass no longer care for messaging anymore.
Another change in the same commit introduced a new second paramter to t/pass, 'chip override'.
Setting it to true tells to robot to chip the ball, even if t/pass wouldn't want to.

However, te/t/volley was not changed. This current commit fixes this.
In a way, this is a simple fixup to 932cf1822857d4e5ca0fe71b75e73ca1749e9207, just over 5 years later.