marvin: te/t/volley: adapt (but worsen) to new dribble tracking
te/t/volley relied on no ball-robot interactions being done in the tracking:
It assumed that a volley-pass would never slow down to velocity 0,
as the tracking would not understand any ball-robot interaction and would project
the ball to pass straight through the robot.

The new tracking does understand that the robot and ball are going to collide.
It sets the velocity of the ball to 0 for the short time
that the _predicted_ world state would have the ball & robot interseect.

This breaks the assumptions and resets the measurement.
The commit fixes that by only ending the measurement once the ball is slow
AND a minimal distance away from the robot.

Note that this solution is stricty worse than the old one:
If we for some reason decide to stop the ball, and shot via stop->stationary
instead of volley, the measurement will still be running and the results will be not helpful at all.

But at least currently, I do not know of a good way to fix this without having to resort to the old
tracking and ra - an obviously not sensible step to take.