glados: a/a/shoot: dont trigger redecide imminent if the ball is moving away from us
We don't redecide on the shoot state if the ball will be reaching our dribbler very soon.
For that, we use the checkedBallRollTime from observer/physics. This function returns
negative infinity, if the ball is actually moving away from the dribbler. This can for
example happen, if we stop the ball, have to move around it first and accidentally bump
the ball away.

Therefore, if we get any negative number from checkedBallRollTime, we assume that the ball
is moving away from us and we don't trigger the imminent criterion.

There is one caveat to this: The dribblefilter in the tracking will set the ball velocity
to zero when the ball hits the robot. In that case, checkedBallRollTime will also return
negative infinity. Therefore, we still want to trigger the imminent criterion, if the ball
is very close to the dribbler pos, even if checkedBallRollTime is negative.