glados: t/m/tigerstestgoalshot: Reliably shoot goals by rotating in front of the defense area
During ShubertOpen, we noticed that TIGERs can shoot goals against us by rotating in front of
our defense area. This is the case because predictShot will first assume that the robot will
shoot in the first corner that is encountered in the rotation. Then the keeper is sent there
with full speed. If however the opponent keeps rotating, predictShot correctly sweeps over the
whole goal to the other corner. The keeper can't keep up with the sudden change of direction
and a goal is easily scored.

The test move TIGERsTestGoalShot aims at replicating this problem. Just place the ball in front
of the defense area, and enable perfect dribbling in 1 vs 1 situations.