glados: trainer/defense.ts: ignore blocked ball trajectories
Ignore the intersections with the defense of the current ball trajectory if we
can be reasonably sure that an opponent robot stops the ball and does not move
out of the way.
Previously our robots sometimes moved to the edge of the defense area even
though the ball never reached it, because it was a pass that an opponent robot
was already waiting for.
This does not handle volley passes, but in the context of the defense this
should not really happen anyways, because it is impossible that the current
trajectory of the ball intersects the defense area and a volley shot can hit
the goal.