glados: o/goal: Add metric for rotating shots
Adds a metric which can be used to identify situtations in which predictShot is rotating.

The setup that was introduced in commit b32669bc8 can be used to reliably score goals
against our defense. TIGERs for example is making heavy use of this tactic.

In order to counter this, we need to be able to detect if the direction given by predictShot
is rotating in a steady, correlated manner. The detection mechanism works as follows:

The directions given by predictShot are stored in a cyclic buffer.
For each pair of consecutive vectors in the buffer, we compute the cross-product between them.
This essentially is a signed version of the angle between the Vectors. For a correlated rotation
the contributions will add up constructively. If predictShot flickers around some mean direction,
the contributions will cancel each other.

Therefore, this so-called rotationConditionNumber, which is by construction between 0 and 1, can
be used to detect correlated rotating motion. Empirically 1e-2 has been shown to be a good
threshold. For smaller numbers, the shot is probably imminent, for larger numbers, the robot is
still rotating.

Regarding the change to the isShot assignment in predictShot:
If isShot is immediately set to true here, the keeper cannot effectively make use of the rotation-
ConditionNumber. This is because we definetly need to defend a shot ball by moving directly into
the predictShot line. However, we can assume that most teams cannot effectively shoot while rotating.
Therefore, as long as the rotationConditionNumber is large enough, it should be safe to not consider
the ball as an instant threat.
This change only affects the following other things in the codebase:
- u/defense calculateBallPosField. This is used in a/d/default, where the dir argument is passed
  to the CenterBack as the defaultPos, but it is never used there. Other than dir, nothing is affected
  by isShot.
- tr/centerback: The defaultCenterback also calls calculateBallPosField. There, isShot not being set
  will cause the CB to defend the center of the goal. Since it is fundamentally impossible to predict
  where exactly the opponent will shoot as long as it is still rotating, defending the center is
  actually the best action in most cases.