glados: t/k/keeper: Only drive as far as necessary while opponent is rotating
In order to defend the setup that was described in commit b32669bc8, the keeper is not allowed
to drive too far into the corner of the goal as long the opponent is still rotating.

This commit introduces the following behavior: The keeper calculates from a crude approximation
how far it needs to be in the corner of the goal to defend a ball that is shot into the corner
on each side. The moveTo position is then bounded into the interval given by these two positions
as long as the rotationConditionNumber is large enough.

In a lot of situations this means that we have a lot less speed in the 'wrong' direction and
don't need to break as much when chasing after predictShot. Additionally, we aren't positioned
as far away from the eventual shot in the first place.

This has been tested to be able to defend against shots into both corners, i.e. both the one
that is sweeped first and the one that is sweeped last by the rotating predictShot.
Fixes #419