attackRatio: refactor attackRatio into enum member type
This splits the attackRatio into three different cases that were previously
already there, but only handled implicitly, which caused them to not work.
The cases are:
  1. Scalable: we have just a rough estimate of the attackers we want to have
    and it is not important to have a specific count. These can therefore be
    trivially scaled when we don't have all 11 robots.
  2. Constant attacker count: we only want an exact number of attackers and
    every other robot should be a defender.
  3. Constant defender count: we want an exact number of defenders and every
    other robot should be an attacker.

This also allows us to actually assure that we have two robots when a pass
happens.
Previously the attack ratio was increased to 2 if it was less, but because of
rounding this does not guarantee that we actually have 2 attackers.