a/a/shoot: do not accept passes if MA is only possible attacker
Previously this was one instance of the infamous "Trying to remove with force
keeping in pool" errors.
This happened, because a defender e.g. a Piggy suggested a pass, but we only
allowed one attacker (the main attacker that had force keeping in pool on,
precisely because it was the main attacker), which made it impossible for the
defender to become an attacker.
That in turn made it impossible to actually accept the pass that was suggested.