glados: t/k/keeper: Experimental fix for overshooting
With a specific setup it was possible to consistently get the keeper to overshoot.
To do this, the ball should be placed 2.75m away from the halfwayline and the
TIGERsTestGoalShot should be triggered such that the shooting robot has to rotate a
little bit.

The underlying problem is probably the expSlowdownTime in robotTimeToPos, which
provides a systematic error such that we will be too early. In close call situations
we dont have time to correct this and the ball passes the keeper while it is over-
shooting.

However, we can check with simple heuristics if we will be early and brake to
avoid this. Since this does not trigger if the timing is very critical, it
should not break anything.