glados: t/a/catchball: Consistently use endSpeed zero
In catchBall the timing is updated at least every 20 frames with robotTimeToBall.
However, in that specific call, we passed the robots maxSpeed as the endSpeed
argument. Since all other timing considerations in catchBall and the call to CMA
all assume we want to catch the ball without final velocity, this is wrong.

This lead to a 20-frame flickering in catchBall in the SchubertOpen final against
TIGERs (1st half, around 1:07 left on the clock). Every 20 frames the moveDest would
be reset to being way closer to the ball because we assume we can arrive with
endSpeed. Then in the subsequent 20 frames, the moveDest is being corrected again.