glados: o/physics: RTTB always sample intersection with current movement direction
In cases where the robot already has a lot of speed orthogonal to the
direction of the ball, robotTimeToPos becomes very sensitive to offsets
orthogonal to the current movement direction.

There might not be a sample point in rttbQuadraticSampling which is
sufficiently close to the current movement direction, such that a
good timing for the robot is achieved. This might lead us to the wrong
conclusion that the robot is not able to reach the Ball, even with
allowed endSpeed.

This commit does not fully solve the problem, since robotTimeForBallTime
does some offsets on the robot position, which are not easy to invert.
Therefore, we only get a sampling position, which is close to being
in the movement direction. In many cases, this is already good enough.

In any case it is better than before and this commit only introduces
one additional sampling point, so it should not break anything else.