glados: t/d/breakpass: Allow for non-orthogonal breakPos
Previously we only considered the orthogonal projection of the robot
position on the ball line as a possible breakPos. However, this might
actually not be the best way to break the pass when we already have
a lot of speed in a slightly different direction.

To avoid the same overshooting problem that was adressed for the
keeper in a426b581, the same workaround is employed and the robot
is forced to brake if it arrives early.

Also, since we do not need to stop the Ball, arriving a little bit
late is not a problem, since the robot has a finite extension.