glados: t/a/catchball: calculateHitTime no longer assumes stationary robot
The previous code for calculateHitTime assumed that the robot is either
stationary or moving towards the ball. This is however not the case in
many situations.

The robot is only going to be actually hit by the ball, if it is able
to kill the orthogonal movement in time before the ball arrives.

The current code produces flickering if the robot has to move through the
ball movement line (e.g. because it already has a lot of transverse speed).
While the robot is in the line, calculateHitTime will assume it is going
to be hit and the robot brakes in order to reach the intersection point.
As soon as it leaves the line again, calculateHitTime returns Infinity
again and the robot again tries to get to the old catchPosition.