glados: t/attackratio: do not use dummy for attack ratio computation
Previously every robot was assumed to be able to become attackers, which was
not that far off when only the keeper could not be an attacker or defender
(and this was recently changed, so the keeper is not counted anymore),
but since dummies were introduced this can be extremely wrong, resulting in
too few defenders.
e.g.:
Let's say we have 2 dummies and 10 robots total and a scalable attack ratio of
7. Previously this would compute floor((7 / 10) * 9) = 6. The 9 is because the kicker
is not counted. This means we have 9 - 6 - 2 = 1 defender, which is VERY BAD,
considering attack ratio of 7 ideally means we have 7 attackers, 1 keeper and
3 defenders.
Now the computation is floor((7 / 10) * 7) = 4, which means we would have
9 - 4 - 2 = 3 defenders.
This is way more in line with what the attack ratio should do.