glados: t/attackratio: change max assumed amount of robots to not include keeper
It never really made sense to use the maximum amount of robots we can have on
the field, since we always want to have a keeper.
The current Division A rules make this especially convenient, since this means
our attackratio is now just a value between 0 and 10, which makes the ratios
easy to think about since an attack ratio of 6 means we want 60% of our robots
to be attackers.
This is also cemented in the type system to make sure no one accidentally
uses a ratio that does not make sense.