glados: a/b/objective: shift zone center towards mean pos of robots
The zones are computed every time the main attacker changes and are usually
clustered around the main attacker, but if we play a long pass, e.g. through the
defense area, almost every single one of our robots tries to move across the
whole field to get to their newly assigned zone.
Now the cluster center is shifted a bit towards the mean pos of all the
attackers, which still fulfills our goal of having the zones close to the
main attacker while also having them close most of where the robots already are.