glados: t/k/keeper: Be more careful regarding possible spin
We do not chase after shots which won't hit our goal
(see b5985757c3a267eefe for more info)

However, the ball might have (a lot) of spin, which means
it might be still dangerous if the ball barely misses the
goal.

This commit disables the keeper bounds for such shots.