glados: u/zone: reuse old default pos if zone is the same
The issue with changing the default pos even though the zone is the same is
that if the zone get recomputed a robot that was almost at the default pos in
its zone might get assigned another zone.
This lead to more time taken up by robots moving into their zones, because we
basically increased the distances our robots need to move.
This prevents that which makes recomputing zones a lot more stable.