glados: objectives: Only construct new MARunner when running objectives
The statements of the form
objective.getMaRunner(agent).agent() === agent
always returned true in the previous implementation, since the
getMaRunner-method always resets the MaRunner if the agent
provided is different.

What we want instead is for the MA to reset the MA-runner
when it is running the objective only once. Then the
checks of the type given above are actually useful.

However, the one occurence in selectObjective was wrong
anyways, as we want to be able to reuse objectives from
different robots.