glados: u/defense: Don't ManMark the opponent CBs
ManMarking the opponent CBs is stupid, because the ManMarkPos
lies a good 1.5m away from our defenseArea, so the ManMarker
would just stand there completely passively.

It is indeed better to be ManMarker than CB if the opponent
CB is actually doing the shot, so we only want to change the
action of the ManMarker if the Ball is in our field.

Just increasing the MarkingDistance will lead to the ManMarkPos
being so far into our defenseArea such that we will just become
CB instead. The cBBackOffDist is fitted in such a way that the
ManMarker will become CB if the target opponent is close the line
that extends the DefenseArea parallel to the HalfWayLine.