glados: t/attackratio: fix trying to remove with force keeping in pool for main/passive
Previously handleBall tried to become main attacker even though we did not allow
any attackers, e.g. in main/passive.
The onlySingleAttackers message is now the limitedAttackerCount message and has
the number of robots as payload.
The case of onlySingleAttackers is now
  limitedAttackerCount != undefined && limitedAttackerCount <= 1
but we now can additionally check in handleBall if limitedAttackerCount === 0.