HACK amun: processor/radiosystem: Let USBThread live longer
This was the case prior to a8f0986ccd1b.

There is a race condition somehwere, where turning radio on or
off lets Ra segfault. It's probably a use-after-free since it segfaults
on a emit. This commit only masks the problem and lets the race conditon
occur less often. However, its better than nothing for now.