glados: t/a/catchball: avoid driving into defenseArea
While it is technically allowed to touch the ball when
we are not completely inside our defense area, imperfect
driving or non-zero endspeed at the target location might
still lead to a multiple defenders foul.

Even if control would be perfect, the rules state that we
should 'keep best effort' to stay outside the defense area.
Therefore this change should be made either way.