glados: a/k/handleball: Add hysteresis to prevent agressiveKeeper
If the keeper tries to pass the ball and the IR-sensor does not trigger
correctly, then we push the ball with our robot. Since the ball is being
accelerated that way, it is going to be detected as 'ball is going to
leave the defense area'. This causes the agressive keeper to trigger, which
in turn causes the robot to rotate, further loosing control over the ball.

We need to account for the fact, that our robots might push the ball quite
some distance before actually shooting them. That means that as long as we
have control over the ball, we want to stay in the branch of the code that
does the pass, movetostaticball or chipaway tasks.