glados: t/a/shoot: Don't assume the ball will hit the dribbler centrally
Due to non-perfect control, robots will often overshoot the target position
by a little bit. However, if we assume that an imminent ball will hit the
dribbler centrally, then the moveTo will also drift together with the
overshooting robot. This renders the problem even worse as now the setpoint
for the control also shifts with the robot, making it even harder to stop.