glados: o/goal: Improve keeper against fast dribbling robots
Before this commit we would assume that a fast ball automatically means that
the ball has been shot. However, this allows teams that can dribble the ball
while moving fast to confuse the keeper very easily.

This commit fixes the issue by not considering the ball to be shot if it is
still very close the the dribbler of the dribbling robot.