glados: a/at/feintgoalshot: Implement feintGoalShot
This is a first try to implement a feint goalshot.
The attacker tries to confuse the keeper by looking at a corner of the goal
and once the keeper goes there to defend the incoming shot, it looks at the
other corner and shoots.

The behavior is tuned in a way that it only turns on very conservatively and
aborts if we are in danger of loosing the ball soon.

This implementation is still rather experimental, and should be improved upon
in the future. This could go together with a possible rework of the shootGoal
It could go together with a rework of the shootgoal task.