seshat: seqlogfilreader: No logspam for qUncompress
This is technically just an output change: The empty group would not have been modified by qUncompress
and a line after that we check for that and abort, it is nicer to avoid calling qUncompress all together for empty m_currentGroups.

It is unclear however why we even have these sections only consisting of invalid packets.