amun/processor: improve computation of division by additionally using allowed number of robots
Previously the division was only computed by using the field size, but it is
less error prone to use the information sent by the referee to compute it.
Specifically the currently allowed number of robots and the number of current
yellow and red cards.
If this information is not present we fall back to the field size computation.
(We do not receive the devision explicitly, so we have to compute it somehow)