core/rng & tests/alphatimetrajectory: Sequence calls to rng
This is an easy mistake to make: I just want two random numbers, so use two calls to rng to get them.
However, as any good PRNG would, core/rng.h has state that updates when a random number is created.
So from a pure C++ point of view, a call to rng ALWAYS HAS A SIDE EFFECT.

Now, look at https://en.cppreference.com/w/cpp/language/eval_order.
With c++17, this is rule 15.
"In a function call, value computations and side effects of the initialization of every parameter are
indeterminately sequenced with respect to value computations and side effects of any other parameter."

For 'indeterminately sequences' they write the following definition:
" Evaluations of A and B are indeterminately sequenced:
  they may be performed in any order but may not overlap:
  either A will be complete before B, or B will be complete before A.
  The order may be the opposite the next time the same expression is evaluated."

Soo - it's technically not a bug if you 'just' care for random numbers.
It'll create random numbers. Without UB. That's also why our extensive tests for core/rng.h
didn't show anything.

But it is a bug if you want to be java: To not only create random numbers on each OS,
but the create the SAME random numbers if seeded, independed of the Toolchain.
And sadly, this is EXACTLY what we want in our tests.

So I changed the two places where we called to rng in an unsequenced way for this test.
There might be more out there, but outside of tests I don't really care. Again,
as I said, it's not a bug. It's not UB. It's just different random numbers on different
Toolchains.

Sadly (or luckely), this does not fix the underlying issue: Now the test fails
on any toolchain, not just clang. Seems to be an issue inside of Alphatimetrajectory to me.

Just in case we want the 'other' sequencing, I also added a specific testcase
'evilTrajectory' that works without an RNG. It should also be easier to debug imo.