ra: teamwidget: fix filter in open file dialog
For some unknown reason, the filter works if it's just one
single file name (*optionally* containing a glob), or a space
separated list of file names. In the latter case however, it seems
the globs are *not* optional, but required. This isn't documented
anywhere, as the filter format is only described by a couple of
examples in the docs.

This commit adds another filter containing a glob, which fixes
the file filtering.