base: remove table.ts
This was an extension of lua tables, which was never really ported and does not
seem to make sense to actually port to typescript as it is an extension of a
construct that does not exist in the language.
Also no one seemed to miss it in the last 5 years where it was not correctly
ported.