glados: o/error: move referee last stop observer to o/referee
This turned out to be more complicated than anticipated, as
isLeavingStop went from a private function that was called
before updateRefereeState, to a public function that was
called after updateRefereeState.

Because of that change, it was necessary to do some more work
to ensure we don't change the behaviour of a public function.

There are three functions whose behaviour we mustn't change:
both of the exported functions `getLastStopTime` and `getLastRefChange`,
as well as `isLeavingStop`, which wasn't exported in o/error,
but needs to be exported from o/referee as it is used in o/error.

`getLastRefChange` simply returns the world time of the first frame
in which the referee had its current state.
`getLastStopTime` returns the world time of the first frame in which
the referee was not in Stop.
`isLeavingStop` returns true if the referee state was Stop in the
previous frame, but isn't Stop in the current frame.