glados: o/referee: write observer to get the last time the referee was in a certain state
This observer already existed for the Stop state.

This commit also slightly changes the semantics of
`getLastTimeInState`, as previously, if the referee was in stop
`getLastStopTime` would return the last time we _left_ stop.
Now we return the current world time. But this isn't a problem since
the only use of this observer can only be reached if the referee is
not in Stop, so this edge chase doesn't make a difference.