ra/robotwidget: remove battery smoothing
The smoothed status of the battery may have had its advantages in the past,
but at this point it causes a lot of confusion.
Without smoothing we are more reliant on a stable return channel, but without
that Ra would also just work on non-existant data for most of its time,
making the UI information equally useless in that case.
If the measurements of the battery status are itself noisy then that should be
filtered on the robot, where all of the necessary/helpful information is and
not in Ra.