data: strip-to-minimal.sh: set commiter and author in the initial commit
Previously, we set the --author option when making the initial commit in
the stripped repo. That works, as long as you have a global name and
email configured in git. But if you dont, git commit will fail, as it
doesnt have a commiter name/email.

To avoid this, instead of overriding the author, we override user.name
and user.email in the commit step.