glados: g/moves: allow MIN_ROBOTS == 0 in moves
This was the intended behavior, and changed (maybe erroneously?)
during porting.

This way, a move can force that there are only defenders, by returning
an empty assignments object. This behavior is also more in line with
the functionality, that a move can simply drop a robot by not assigning
it any task, as long as the number of assigned robots is still between
MIN_ROBOTS and MAX_ROBOTS.

This also allows us to remove a wrong and unnecessary assertion that
was checking if the number of attackers is zero.