amun: trajectorypath: make calculatePosition behave equivalent to calculateTrajectory
calculatePosition is essentially a more efficient version of
calculateTrajectory that only computes the end position.

In the general case, that invariant was correct. But in commit
4d66e7f57405fe40cf057413b14ef6b5c723f6a7, a special case for short times
was added in calculateTrajectory, but not in calculatePosition.

This lead to findTrajectory returning wrong trajectories, as it thought
the end position (returned by calculatePosition) was close to the
target, but the resulting trajectory (returned by calculateTrajectory)
did not end there.