cpptests: alphatimetrajectory: remove evilTrajectory test
The test was a certain iteration of the findTrajectory test
which generated a wrong trajectory with a too high acceleration.

The previous commit (c407b8dd9ba50c875921ae499ab8c05520988953)
fixed that issue, and now the test failed because findTrajectory
doesnt find any trajectory, which is fine as this is sometimes the case.