amun: trajectorypath: adjust endspeed when endspeedtype is fast
There are two possible endspeed types: fast and exact. Exact means
the endspeed should be matched exactly, and fast means the x and y
components of the actual endspeed should only be as the respective
components of the endspeed parameter.

The minTime of a trajectory is the time necessary to go from v0 to
v1, so ||v1 - v0|| / acc.

When generating a minTimeTrajectory, it expects the endspeed to be
reachable in minTime. This requires the endspeed to be adjusted with
minTimeEndSpeed. This step was missing before the call to
minTimeTrajectory.