cmake: add some target link dependencies explicitly with add_dependencies
No idea why, but cmake apparently is not smart enough to figure out, that if
you want to link with a target that you're building (e.g. bullet for the
simulator) that it should build that target first.
When building with multiple jobs it usually just worked before already, but
sometimes you had to do multiple runs until the race conditions line up
correctly.
When building with one thread it just reproducibly failed to build.
This should now work(tm).