cmake: set protobuf_generate_DEPENDENCIES when building protobuf
If this is not set cmake does not wait until protoc is built to start
generating the cpp and header files.
When building with multiple threads protoc was usually done building by
the time cmake decided to use it, but relying on race conditions in the build
systems is not the best idea and with one thread it reproducibly failed every
time.