cmake: fix bug where cmake config fails if protobuf is not installed
The if condition contains a check if ${Protobuf_VERSION} is too new,
but if protobuf was not found at all this variable is empty, causing cmake
to die, because of course it can't handle this. Like, When would it ever happen
that a library, that is searched for, is both checked if it exists AND which
version it has.