glados: te/m/race: add random offset to positions to avoid wearing down the carpet
When testing a lot with race, the tracks the robot drive become very
visible in the carpet as it wears down. This happens in the areas were
the robots break and accelerate, as well as in the straight line
segments between the start and the end.

This commit adds a small random offset in the x and y coordinates so
that the wear of the carpet is spread over a larger area, and is that
way more uniform and less localized.