amun/tracking: remove setting of unused H matrix for future kalman
Commit 5338a7223a2161fefd9beefeecbed66ab3ec5772 changed the way we handle radio
commands in the tracker. Previously they were used as a kind of measurement, so
the measurement matrix H of the future kalman had to be set to specific values
distinct from the H of the present kalman, but now they are used as a control
input into the system, so the H matrix of the future kalman is just unused and
we assume that it was just forgotten to remove this part.