amun/js_amun: send all radio responses to strategy regardless of if they are visible in vision
This was previously avoided as a performance optimization, but we plan to rely
way more on radio responses in the future and for reasons like testing off the
field how errors behave it is reasonable to just copy all of the responses.
Additionally of all the fields that have been avoided to send to the strategy
this seems to be the smallest.